float -> double

+++ src/modules/bifrost/geometry/BifrostGeometry.cpp
@@ -63,7 +63,7 @@ int BifrostGeometry::yOffset(int TubeId) {
 }
 
 
-std::pair<int, float> BifrostGeometry::calcUnitAndPos(int Group, int AmpA, int AmpB) {
+std::pair<int, double> BifrostGeometry::calcUnitAndPos(int Group, int AmpA, int AmpB) {
 
   if (AmpA + AmpB == 0) {
     XTRACE(DATA, DEB, "Sum of amplitudes is 0");
@@ -71,7 +71,7 @@ std::pair<int, float> BifrostGeometry::calcUnitAndPos(int Group, int AmpA, int A
     return InvalidPos;
   }
 
-  float GlobalPos = 1.0 * AmpA / (AmpA + AmpB); // [0.0 ; 1.0]
+  double GlobalPos = 1.0 * AmpA / (AmpA + AmpB); // [0.0 ; 1.0]
   if ((GlobalPos < 0) or (GlobalPos > 1.0)) {
     XTRACE(DATA, WAR, "Pos %f not in unit interval", GlobalPos);
     return InvalidPos;
@@ -90,7 +90,7 @@ std::pair<int, float> BifrostGeometry::calcUnitAndPos(int Group, int AmpA, int A
   auto & Intervals = CaenCDCalibration.Intervals[Group];
   double Lower = Intervals[Unit].first;
   double Upper = Intervals[Unit].second;
-  float RawUnitPos = (GlobalPos - Lower)/(Upper - Lower);
+  double RawUnitPos = (GlobalPos - Lower)/(Upper - Lower);
 
   XTRACE(DATA, DEB, "Unit %d, GlobalPos %f, RawUnitPos %f", Unit, GlobalPos, RawUnitPos);
   return std::make_pair(Unit, RawUnitPos);
@@ -102,7 +102,7 @@ uint32_t BifrostGeometry::calcPixel(DataParser::CaenReadout &Data) {
   int yoff = yOffset(Data.TubeId);
 
   int Group = Data.RingId * TripletsPerRing + Data.TubeId;
-  std::pair<int, float> UnitPos = calcUnitAndPos(Group, Data.AmpA, Data.AmpB);
+  std::pair<int, double> UnitPos = calcUnitAndPos(Group, Data.AmpA, Data.AmpB);
 
   if (UnitPos.first == -1) {
     return 0;
